WalletConnect on xDai
Added some comments around the web3Modal pertaining to apps deployed to xDai network. If these are not added, and app.jsx/hardhat.config ARE pointed to xDai. Then when you connect using WalletConnect it wants to make all transactions on mainnet ETH. Even when you change wallet settings to be on the xDai chain. In this changes, when you uncomment the code above, wallets like AlphaWallet when on xDai chain can make transactions on your app right from your mobile phone! This opens up more options for an xDai dev not just MetaMask.